Add files via upload
The anonymous inner class is created and used in Lines 17–22. It monitors keyboard input using
the only method in the KeyListener interface the application requires and displays that input
by updating the keyLabel instance variable.
The listener is created anonymously with the call to new KeyAdapter() followed by the
definition of the class. The class overrides the keyTyped() method so that when a key is
pressed, it is retrieved with a call to getKeyChar() and displayed by setting the value of
keyLabel, which is a JLabel component.
The anonymous inner class does something that a normal helper class could not do: access the
keyLabel instance variable. That variable belongs to the KeyViewer class. Inner classes can
access the methods and variables of their enclosing class
Anonymous inner classes cannot define constructors, so they are more limited than a nonanonymous inner class.

Anonymous Inner Classes
A common task you undertake in Java programming is to create a class that will be used only
once and has a simple purpose. A special kind of inner class is perfect for this purpose.
An anonymous inner class is one that has no name and is declared and created at the same
time.
To use one, you replace a reference to an object with the new keyword, a call to a constructor,
and the class definition inside { and } characters.
Here’s some code that does not use an anonymous inner class:
WorkerClass worker = new WorkerClass();
Thread main = new Thread(worker);
main.start();
The worker object presumably implements the Runnable interface and can be run as a thread.
If the code in WorkerClass is short and simple, and the class only needs to be used once, it
could be worthwhile to put it in an anonymous inner class. Here’s a new version to do that:
Thread main = new Thread(new Runnable() {
 public void run() {
 // thread's work to perform goes here
 }
});
main.start();
226 HOUR 16: Using Inner Classes and Closures
The anonymous inner class has replaced the reference to worker with the following code:
new Runnable() {
 public void run() {
 // thread's work to perform goes here
 }
}
This creates an anonymous class that implements the Runnable interface and overrides the
run() method. The statements inside the method would perform whatever work the class
requires.
This concept will be easier to comprehend with a full example of how an anonymous inner class
is written and why it’s useful.